requirements.txt
Django==1.6.2
South==0.8.4
dj-database-url==0.3.0
dj-static==0.0.5
django-tastypie==0.11.0
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
rauth==0.7.0
requests==2.2.1
six==1.6.1
static==1.0.2
